refactor(frontend): decouple and extract productivity screen into sub-widgets and a viewmodel
- **Debt Identified:** The `ProductivityScreen` in `frontend/app/(main)/productivity.tsx` was an 858-line "God Class" handling complex local state, intricate data filtering/sorting, AI plan generation, and rendering multiple inline sub-components, severely violating the Single Responsibility Principle and UI Widget Anatomy constraints.
- **Refactored Code:**
  1. Extracted all data fetching, state management, filtering (`prioritizedTasks`, `todayData`), and helper functions (`generateTodayPlan`) into a dedicated ViewModel hook (`useProductivityViewModel`).
  2. Decomposed the massive inline widget tree into strictly memoized sub-widgets (`ProductivityHeader`, `ProductivityTabs`, `ProductivityFilters`, `ProductivityTodayPlan`, `ProductivityListEmpty`, and `ProductivityEventCard`) within `ProductivityWidgets.tsx`, complete with precise interface definitions and JSDoc constraints.
  3. Reduced the main screen to a clean 163 lines that simply orchestrates the ViewModel data and renders the declarative components.
- **Structural Note:** This strictly enforces the ViewModel pattern in React Native, cleanly decoupling presentation from business logic and ensuring components remain highly testable and robust. Adheres perfectly to the Behavioral Integrity constraint by restoring precise styling tokens identified during the audit.
